requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1617740417847/work
boto3==1.17.23
botocore==1.20.23
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django @ file:///tmp/build/80754af9/django_1617740316746/work
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-storages==1.11.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
Pillow==8.1.0
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
python-dateutil==2.8.1
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
urllib3==1.26.3
whitenoise @ file:///home/conda/feedstock_root/build_artifacts/whitenoise_1602596638218/work
wincertstore==0.2
